source/status/requirements.txt
# status/requirements.txt
# Список зависимостей Python для Flask-приложения

Flask==2.3.3
psycopg2-binary==2.9.9 # Для работы с PostgreSQL
gunicorn==21.2.0       # WSGI сервер для production
python-dateutil==2.8.2 # Для парсинга дат ISO 8601 от PowerShell
Flask-SocketIO==5.3.6  # <<< Добавлено для WebSockets

# Выберите один из асинхронных режимов и раскомментируйте:
eventlet==0.33.3    # <<< Для async_mode='eventlet'
# gevent==23.9.1      # <<< Для async_mode='gevent'
# gevent-websocket==0.10.1 # <<< Зависимость для gevent

Flask-Login==0.6.3  # <<< ДОБАВЛЕНО
Werkzeug>=2.3.0     # <<< УБЕДИТЕСЬ, что Werkzeug достаточно новый для хеширования